update AuthDtos and AuthService to use Integer for phmAreaId and integrate PhmAreaRepository

+++ src/main/java/com/moh/moh_backend/dto/AuthDtos.java
@@ -34,7 +34,7 @@ public class AuthDtos {
         }
         public static class MidwifeDetails {
             public String name;
-            public Long phmAreaId;
+            public Integer phmAreaId;
             public String contactNumber;
             public String email;
             public String assignmentDate;
